Move test helper classes into Infrastructure namespace
