+++ src/libraries/System.IO.Pipelines/tests/StreamPipeReaderCopyToAsyncTests.cs
@@ -339,68 +339,5 @@ namespace System.IO.Pipelines.Tests
             Assert.True(startPosition.Equals(wrappedPipeReader.LastConsumed));
             Assert.True(startPosition.Equals(wrappedPipeReader.LastExamined));
         }
-
-        private class ThrowingStream : ThrowAfterNWritesStream
-        {
-            public ThrowingStream() : base(0)
-            {
-            }
-        }
-
-        private class TestPipeReader : PipeReader
-        {
-            private readonly PipeReader _inner;
-
-
-            public TestPipeReader(PipeReader inner)
-            {
-                _inner = inner;
-            }
-
-            public ReadResult LastReadResult { get; private set; }
-            public SequencePosition LastConsumed { get; private set; }
-            public SequencePosition LastExamined { get; private set; }
-
-            public override void AdvanceTo(SequencePosition consumed)
-            {
-                LastConsumed = consumed;
-                LastExamined = consumed;
-                _inner.AdvanceTo(consumed);
-            }
-
-            public override void AdvanceTo(SequencePosition consumed, SequencePosition examined)
-            {
-                LastConsumed = consumed;
-                LastExamined = examined;
-                _inner.AdvanceTo(consumed);
-            }
-
-            public override void CancelPendingRead()
-            {
-                _inner.CancelPendingRead();
-            }
-
-            public override void Complete(Exception exception = null)
-            {
-                _inner.Complete(exception);
-            }
-
-            public override async ValueTask<ReadResult> ReadAsync(CancellationToken cancellationToken = default)
-            {
-                LastReadResult = await _inner.ReadAsync(cancellationToken);
-                return LastReadResult;
-            }
-
-            public override bool TryRead(out ReadResult result)
-            {
-                if (_inner.TryRead(out result))
-                {
-                    LastReadResult = result;
-                    return true;
-                }
-
-                return false;
-            }
-        }
     }
 }
