+++ src/libraries/System.IO.Pipelines/tests/System.IO.Pipelines.Tests.csproj
@@ -8,6 +8,8 @@
     <Compile Include="Infrastructure\HeapBufferPool.cs" />
     <Compile Include="Infrastructure\DisposeTrackingBufferPool.cs" />
     <Compile Include="Infrastructure\ReadOnlyStream.cs" />
+    <Compile Include="Infrastructure\TestPipeReader.cs" />
+    <Compile Include="Infrastructure\ThrowingStream.cs" />
     <Compile Include="Infrastructure\WriteCheckStream.cs" />
     <Compile Include="Infrastructure\WriteOnlyStream.cs" />
     <Compile Include="Infrastructure\CancelledWritesStream.cs" />
